agent, common: move validateNetworkId to indexer-common
